Update README.md
Sentiment Analysis on IMDB Movie Reviews — Using fine-tuning BERT
Developed sentiment analysis system for movie reviews using fine-tuned BERT, achieving 89% accuracy with efficient GPU utilization, 
implemented custom classifier layers, trained model with Adam optimizer, reducing validation loss by 59%, and improving NLP accuracy by 97% in 2 epochs
compared to baseline models.

The main goal is to create a model that can accurately predict the sentiment of sentences, and the steps taken align with best practices for working with BERT-based models in natural language processing tasks.
Here are some potential reasons for fine-tuning this model for Sentiment Analysis:

Preprocessing for BERT Input:
BERT expects input data in a specific format, including tokenization, special tokens, and attention masks. The preprocessing steps ensure that the data is formatted appropriately for input into the BERT model.

Fine-Tuning BERT for Sentiment Analysis:
Pre-training a BERT model on a large corpus gives it a general understanding of language. However, fine-tuning is required to adapt the model to a specific task, such as sentiment analysis. The labeled dataset is used for this fine-tuning process to make the model more domain-specific.

Transfer Learning:
BERT leverages transfer learning, where knowledge gained from a large pre-training task is transferred to a smaller, task-specific dataset. This approach is effective when you have limited labeled data for a specific task (sentiment analysis in this case).

Evaluation on Test Set:
The dataset is split into training and testing sets to evaluate the model's performance on unseen data. This is crucial to assess how well the model generalizes to new, unseen examples and to avoid overfitting to the training data.

Application to New Sentences:
Once the model is trained and evaluated, it can be applied to new sentences to predict their sentiments. This is demonstrated in the last part of the code where two example sentences are given, and the model predicts whether they have a positive or negative sentiment.